requirements.txt
torch>=2.0.0
gymnasium>=0.28.1
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
wandb>=0.13.0
python-nmap>=0.7.1
python-libnmap>=0.7.2
pymetasploit3>=1.0.3
nest-asyncio>=1.5.5
paramiko>=2.11.0
requests>=2.28.0
aiohttp>=3.8.3
transformers>=4.25.0
python-dotenv>=1.0.0 